+++ lib/Model/NoteFieldsResponseAllOf.php
@@ -66,7 +66,7 @@ class NoteFieldsResponseAllOf implements ModelInterface, ArrayAccess, JsonSerial
       * @phpsalm-var array<string, string>
       */
     protected static array $openAPITypes = [
-        'data' => '\Pipedrive\Model\NoteField[]',
+        'data' => '\Pipedrive\Model\GetNoteField[]',
         'additional_data' => '\Pipedrive\Model\AdditionalData'
     ];
 
@@ -237,7 +237,7 @@ class NoteFieldsResponseAllOf implements ModelInterface, ArrayAccess, JsonSerial
     /**
      * Gets data
      *
-     * @return \Pipedrive\Model\NoteField[]|null
+     * @return \Pipedrive\Model\GetNoteField[]|null
      */
     public function getData()
     {
@@ -247,7 +247,7 @@ class NoteFieldsResponseAllOf implements ModelInterface, ArrayAccess, JsonSerial
     /**
      * Sets data
      *
-     * @param \Pipedrive\Model\NoteField[]|null $data data
+     * @param \Pipedrive\Model\GetNoteField[]|null $data data
      *
      * @return self
      */